+++ src/main/scala/edu/rit/cs/mmior/jsonoid/discovery/schemas/AnySchema.scala
@@ -32,4 +32,8 @@ final case class AnySchema(
 
     newSchema
   }
+
+  override def isCompatibleWith(other: JsonSchema[_])(implicit
+      p: JsonoidParams
+  ): Boolean = true
 }