+++ src/main/scala/edu/rit/cs/mmior/jsonoid/discovery/schemas/ZeroSchema.scala
@@ -29,4 +29,9 @@ final case class ZeroSchema(
 
     newSchema
   }
+
+  // Compatibility doesn't really even make sense here, so go with false
+  override def isCompatibleWith(other: JsonSchema[_])(implicit
+      p: JsonoidParams
+  ): Boolean = false
 }